requirements.txt
# LINE Bot 活動管理系統依賴
Flask==2.2.3
line-bot-sdk==3.16.2
python-dotenv==1.0.0
notion-client==2.0.0
APScheduler==3.10.1
requests==2.28.2
requests-toolbelt==1.0.0  # 用於處理Rich Menu圖片上傳
pytz==2023.3
gunicorn==21.2.0
Pillow==10.0.0  # 用於生成Rich Menu圖片 